+++ Angular-Front-End/app-front-end/src/app/app-routing.module.ts
@@ -7,12 +7,14 @@ import { LoginPageComponent } from './components/login-page/login-page.component
 import { SignUpPageComponent } from './components/sign-up-page/sign-up-page.component';
 import { EventsUserComponent } from './components/events-user/events-user.component';
 import { AuthGuard } from './services/authentication/auth-guard.service';
+import { HomePageAfterLoginComponent } from './components/home-page-after-login/home-page-after-login.component';
+import { ContactsComponent } from './components/contacts/contacts.component';
 
 
 const routes: Routes = [
   {
     path: '', 
-    redirectTo: 'homePage',
+    redirectTo: 'welcome',
     pathMatch: 'full'
   },
   {
@@ -20,12 +22,18 @@ const routes: Routes = [
     component: HomeComponent
   },
   {
-    path: "homePage",
+    path: "welcome",
     component: HomePageComponent
   },
   {
+    path: "homePage",
+    component: HomePageAfterLoginComponent,
+    canActivate: [AuthGuard]
+  },
+  {
     path: "communities",
-    component: CommunitiesListComponent
+    component: CommunitiesListComponent,
+    canActivate: [AuthGuard]
   },
   {
     path: "login",
@@ -40,6 +48,10 @@ const routes: Routes = [
     component: EventsUserComponent,
     canActivate: [AuthGuard]
   },
+  {
+    path: "contacts",
+    component: ContactsComponent
+  }
 ];
 
 @NgModule({