+++ Angular-Front-End/app-front-end/src/app/components/home-page/home-page.component.ts
@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { AuthenticationService } from 'src/app/services/authentication/authentication.service';
+import { UserModel } from 'src/app/models/user.model';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-home-page',
@@ -7,9 +10,19 @@ import { Component, OnInit } from '@angular/core';
 })
 export class HomePageComponent implements OnInit {
 
-  constructor() { }
+  currentUser: UserModel;
+
+  constructor(
+    private authenticationService: AuthenticationService,
+    private router: Router
+    ) {
+      this.authenticationService.currentUser.subscribe(x => this.currentUser = x);
+    }
 
   ngOnInit(): void {
+    if(this.currentUser) {
+      this.router.navigate(['homePage']);
+    }
   }
 
 }