+++ Angular-Front-End/app-front-end/src/app/models/event.model.ts
@@ -1,0 +1,10 @@
+import { Identifiable } from './identifiable';
+
+export interface EventModel extends Identifiable{
+    id: string;
+    name: string;
+    start: string;
+    end: string;
+    cID: string;
+    cName: string;
+}