+++ Database/SQLTables.sql
@@ -122,8 +122,8 @@ UNLOCK TABLES;
 -- --------------------Events-------------------------
 LOCK TABLES Events WRITE;
 -- Test Values
-INSERT INTO Events VALUES (1,'Reunião','2020-02-24','2020-02-25',1,'ASC'),
-			  (2,'Almoço_no_LASIGE','2020-03-14','2020-03-14',2,'LASIGE');
+INSERT INTO Events VALUES (1,'Reunião','2020-02-24','2020-02-25',4,'ASC'),
+			  (2,'Almoço_no_LASIGE','2020-03-14','2020-03-14',5,'LASIGE');
 UNLOCK TABLES;
 
 -- --------------RolesUsersCommunities----------------